sv-comp/scripts/requirements.txt
argcomplete>=3.2.2
black>=24.2.0
httpx[http2]>=0.25.0
jsonschema>=4.19.2
lxml>=5.2.1
pandas>=2.2.0
progressbar2>=4.2.0
pyyaml>=6.0.1
yq>=3.1.0
coloredlogs==15.0.1
matplotlib>=3.9.2
xmlstarlet>=1.6.1